Refactor: package plugin related classes together
